feat: revamp module organization
