+++ colpali_engine/loss/__init__.py
@@ -1,1 +1,2 @@
-from .colbert_loss import ColbertLoss
+from bi_encoder_losses import *
+from late_interaction_losses import *
