+++ colpali_engine/models/__init__.py
@@ -1,0 +1,2 @@
+from bi_encoders import *
+from late_interaction import *
