+++ colpali_engine/models/bi_encoders/__init__.py
@@ -1,0 +1,3 @@
+from biidefics_architecture import *
+from bipali_architecture import *
+from bisiglip_architecture import *
