+++ colpali_engine/models/late_interaction/__init__.py
@@ -1,0 +1,3 @@
+from colidefics_architecture import *
+from colpali_architecture import *
+from colsiglip_architecture import *
