+++ colpali_engine/utils/colpali_processing_utils.py
@@ -1,4 +1,4 @@
-# Utils for processing images and queries for ColPaLi
+"""Utils for processing images and queries for ColPaLi"""
 
 
 def process_images(processor, images, max_length: int = 50):
