+++ colpali_engine/utils/wrapper.py
@@ -1,16 +1,7 @@
 import importlib
 
-from colpali_engine.models.clip_baselines import ColSigLIP, SigLIP
-from colpali_engine.models.colbert_architectures import (
-    BiBERT,
-    BiXLMRoBERTa,
-    ColBERT,
-    ColCamembert,
-    ColLlama,
-    ColXLMRoBERTa,
-)
-from colpali_engine.models.idefics_colbert_architecture import BiIdefics, ColIdefics
-from colpali_engine.models.paligemma_colbert_architecture import BiNewSiglip, BiPali, ColNewSiglip, ColPali
+from colpali_engine.models.bi_encoders import BiIdefics, BiNewSiglip, BiPaliLast, BiPaliMean, BiSigLIP
+from colpali_engine.models.late_interaction import ColIdefics, ColNewSiglip, ColPali, ColSigLIP
 
 if importlib.util.find_spec("transformers") is not None:
     from transformers import AutoProcessor, AutoTokenizer
@@ -38,23 +29,13 @@ if importlib.util.find_spec("transformers") is not None:
 
             training_objective = kwargs.pop("training_objective", "colbertv1")
 
-            if "camembert" in pretrained_model_name_or_path:
-                return ColCamembert.from_pretrained(*args, **kwargs)
-            elif "xlm-roberta" in pretrained_model_name_or_path:
-                if training_objective == "biencoder":
-                    return BiXLMRoBERTa.from_pretrained(*args, **kwargs)
-                return ColXLMRoBERTa.from_pretrained(*args, **kwargs)
-            elif (
-                "llama" in pretrained_model_name_or_path.lower() or "croissant" in pretrained_model_name_or_path.lower()
-            ):
-                return ColLlama.from_pretrained(*args, **kwargs)
-            elif "idefics2" in pretrained_model_name_or_path:
+            if "idefics2" in pretrained_model_name_or_path:
                 if training_objective == "biencoder":
                     return BiIdefics.from_pretrained(*args, **kwargs)
                 return ColIdefics.from_pretrained(*args, **kwargs)
             elif "siglip" in pretrained_model_name_or_path:
                 if training_objective == "biencoder_mean":
-                    return SigLIP.from_pretrained(*args, **kwargs)
+                    return BiSigLIP.from_pretrained(*args, **kwargs)
                 elif training_objective == "colbertv1":
                     return ColSigLIP.from_pretrained(*args, **kwargs)
                 else:
@@ -73,9 +54,7 @@ if importlib.util.find_spec("transformers") is not None:
                 else:
                     raise ValueError(f"Training objective {training_objective} not recognized")
             else:
-                if training_objective == "biencoder":
-                    return BiBERT.from_pretrained(*args, **kwargs)
-                return ColBERT.from_pretrained(*args, **kwargs)
+                raise ValueError(f"Model `{pretrained_model_name_or_path}` not supported")
 
 else:
     raise ModuleNotFoundError("Transformers must be loaded")
